# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Skillbytes API - PBL Module",
        "description": "REST API for Problem-Based Learning (PBL) Case Management System",
        "contact": {
            "email": "support@skillbytes.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development"
        },
        {
            "url": "http://145.79.13.180/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/lessons/{lesson}/mark-completed": {
            "post": {
                "tags": [
                    "User Progress"
                ],
                "summary": "Mark lesson as completed",
                "description": "Mark a lesson as completed for the authenticated user.",
                "operationId": "7070f08bd1d9f90b1dd70b63e3548c57",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "Lesson ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "lesson_id": {
                                        "description": "Lesson ID",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lesson marked as completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "lessonId": {
                                            "type": "integer"
                                        },
                                        "completed": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "completedAt": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Lesson not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/my-progress": {
            "get": {
                "tags": [
                    "User Progress"
                ],
                "summary": "Get user's progress",
                "description": "Retrieve the authenticated user's overall progress including completed lessons and mata_pelajarans.",
                "operationId": "e75e84df39c24f90348d8727514b23f0",
                "responses": {
                    "200": {
                        "description": "User progress retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "userId": {
                                            "type": "integer"
                                        },
                                        "totalLessons": {
                                            "type": "integer"
                                        },
                                        "completedLessons": {
                                            "type": "integer"
                                        },
                                        "progressPercentage": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "mataPelajarans": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token for API authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}